+++ src/jsTemplate/events/client/error.js
@@ -1,0 +1,6 @@
+//here the event starts
+module.exports = client => {
+    console.error();
+}
+
+/** Template by Tomato#6966 | https://github.com/Tomato6966/Discord-Js-Handler-Template */