+++ src/jsTemplate/events/client/ready.js
@@ -1,0 +1,30 @@
+//here the event starts
+const config = require("../../botconfig/config.json")
+module.exports = client => {
+  try{
+    const stringlength = 69;
+    console.log("\n")
+    console.log(`     ┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓`.bold.brightGreen)
+    console.log(`     ┃ `.bold.brightGreen + " ".repeat(-1+stringlength-` ┃ `.length)+ "┃".bold.brightGreen)
+    console.log(`     ┃ `.bold.brightGreen + `Discord Bot is online!`.bold.brightGreen + " ".repeat(-1+stringlength-` ┃ `.length-`Discord Bot is online!`.length)+ "┃".bold.brightGreen)
+    console.log(`     ┃ `.bold.brightGreen + ` /--/ ${client.user.tag} /--/ `.bold.brightGreen+ " ".repeat(-1+stringlength-` ┃ `.length-` /--/ ${client.user.tag} /--/ `.length)+ "┃".bold.brightGreen)
+    console.log(`     ┃ `.bold.brightGreen + " ".repeat(-1+stringlength-` ┃ `.length)+ "┃".bold.brightGreen)
+    console.log(`     ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛`.bold.brightGreen)
+  }catch{ /* */ }
+
+  try{
+    client.user.setActivity(client.user.username, { type: "PLAYING" });
+  }catch (e) {
+      console.log(String(e.stack).red);
+  }
+  //Change status each 10 minutes
+  setInterval(()=>{
+    try{
+      client.user.setActivity(client.user.username, { type: "PLAYING" });
+    }catch (e) {
+        console.log(String(e.stack).red);
+    }
+  }, 10*60*1000)
+}
+
+/** Template by Tomato#6966 | https://github.com/Tomato6966/Discord-Js-Handler-Template */