+++ src/jsTemplate/events/guild/message.js
@@ -1,0 +1,128 @@
+/**
+  * @INFO
+  * Loading all needed File Information Parameters
+*/
+const config = require("../../botconfig/config.json"); //loading config file with token and prefix, and settings
+const ee = require("../../botconfig/embed.json"); //Loading all embed settings like color footertext and icon ...
+const Discord = require("discord.js"); //this is the official discord.js wrapper for the Discord Api, which we use!
+const { escapeRegex} = require("../../handlers/functions"); //Loading all needed functions
+//here the event starts
+module.exports = async (client, message) => {
+  try {
+    //if the message is not in a guild (aka in dms), return aka ignore the inputs
+    if (!message.guild) return;
+    // if the message  author is a bot, return aka ignore the inputs
+    if (message.author.bot) return;
+    //if the channel is on partial fetch it
+    if (message.channel.partial) await message.channel.fetch();
+    //if the message is on partial fetch it
+    if (message.partial) await message.fetch();
+    //get the current prefix from the botconfig/config.json
+    let prefix = config.prefix
+    //the prefix can be a Mention of the Bot / The defined Prefix of the Bot
+    const prefixRegex = new RegExp(`^(<@!?${client.user.id}>|${escapeRegex(prefix)})\\s*`);
+    //if its not that then return
+    if (!prefixRegex.test(message.content)) return;
+    //now define the right prefix either ping or not ping
+    const [, matchedPrefix] = message.content.match(prefixRegex);
+    //create the arguments with sliceing of of the rightprefix length
+    const args = message.content.slice(matchedPrefix.length).trim().split(/ +/);
+    //creating the cmd argument by shifting the args by 1
+    const cmd = args.shift().toLowerCase();
+    //if no cmd added return error
+     if (cmd.length === 0){
+      if(matchedPrefix.includes(client.user.id))
+        return message.channel.send(new Discord.MessageEmbed()
+          .setColor(ee.color)
+          .setFooter(ee.footertext,ee.footericon)
+          .setTitle(`Hugh? I got pinged? Imma give you some help`)
+          .setDescription(`To see all Commands type: \`${prefix}help\``)
+        );
+      return;
+      }
+    //get the command from the collection
+    let command = client.commands.get(cmd);
+    //if the command does not exist, try to get it by his alias
+    if (!command) command = client.commands.get(client.aliases.get(cmd));
+    //if the command is now valid
+    if (command){
+        if (!client.cooldowns.has(command.name)) { //if its not in the cooldown, set it too there
+            client.cooldowns.set(command.name, new Discord.Collection());
+        }
+        const now = Date.now(); //get the current time
+        const timestamps = client.cooldowns.get(command.name); //get the timestamp of the last used commands
+        const cooldownAmount = (command.cooldown || 1.5) * 1000; //get the cooldownamount of the command, if there is no cooldown there will be automatically 1 sec cooldown, so you cannot spam it^^
+        if (timestamps.has(message.author.id)) { //if the user is on cooldown
+          const expirationTime = timestamps.get(message.author.id) + cooldownAmount; //get the amount of time he needs to wait until he can run the cmd again
+          if (now < expirationTime) { //if he is still on cooldonw
+            const timeLeft = (expirationTime - now) / 1000; //get the lefttime
+            return message.channel.send(new Discord.MessageEmbed()
+              .setColor(ee.wrongcolor)
+              .setFooter(ee.footertext,ee.footericon)
+              .setTitle(`❌ Please wait ${timeLeft.toFixed(1)} more second(s) before reusing the \`${command.name}\` command.`)
+            ); //send an information message
+          }
+        }
+        timestamps.set(message.author.id, now); //if he is not on cooldown, set it to the cooldown
+        setTimeout(() => timestamps.delete(message.author.id), cooldownAmount); //set a timeout function with the cooldown, so it gets deleted later on again
+      try{
+        //try to delete the message of the user who ran the cmd
+        try{  message.delete();   }catch{}
+        //if Command has specific permission return error
+        if(command.memberpermissions && !message.member.hasPermission(command.memberpermissions)) {
+          return message.channel.send(new Discord.MessageEmbed()
+            .setColor(ee.wrongcolor)
+            .setFooter(ee.footertext, ee.footericon)
+            .setTitle("❌ Error | You are not allowed to run this command!")
+            .setDescription(`You need these Permissions: \`${command.memberpermissions.join("`, ``")}\``)
+          ).then(msg=>msg.delete({timeout: 5000}).catch(e=>console.log("Couldn't Delete --> Ignore".gray)));
+        }
+        //if the Bot has not enough permissions return error
+        let required_perms = ["ADD_REACTIONS","PRIORITY_SPEAKER","VIEW_CHANNEL","SEND_MESSAGES",
+        "EMBED_LINKS","CONNECT","SPEAK","DEAFEN_MEMBERS"]
+        if(!message.guild.me.hasPermission(required_perms)){
+          try{ message.react("❌"); }catch{}
+          return message.channel.send(new Discord.MessageEmbed()
+            .setColor(ee.wrongcolor)
+            .setFooter(ee.footertext, ee.footericon)
+            .setTitle("❌ Error | I don't have enough Permissions!")
+            .setDescription("Please give me just `ADMINISTRATOR`, because I need it to delete Messages, Create Channel and execute all Admin Commands.\n If you don't want to give me them, then those are the exact Permissions which I need: \n> `" + required_perms.join("`, `") +"`")
+          )
+        }
+        //run the command with the parameters:  client, message, args, user, text, prefix,
+        command.run(client, message, args, message.member, args.join(" "), prefix);
+      }catch (e) {
+        console.log(String(e.stack).red)
+        return message.channel.send(new Discord.MessageEmbed()
+          .setColor(ee.wrongcolor)
+          .setFooter(ee.footertext, ee.footericon)
+          .setTitle("❌ Something went wrong while, running the: `" + command.name + "` command")
+          .setDescription(`\`\`\`${e.message}\`\`\``)
+        ).then(msg=>msg.delete({timeout: 5000}).catch(e=>console.log("Couldn't Delete --> Ignore".gray)));
+      }
+    }
+    else //if the command is not found send an info msg
+    return message.channel.send(new Discord.MessageEmbed()
+      .setColor(ee.wrongcolor)
+      .setFooter(ee.footertext, ee.footericon)
+      .setTitle(`❌ Unkown command, try: **\`${prefix}help\`**`)
+      .setDescription(`To play Music simply type \`${prefix}play <Title / Url>\``)
+    ).then(msg=>msg.delete({timeout: 5000}).catch(e=>console.log("Couldn't Delete --> Ignore".gray)));
+  }catch (e){
+    return message.channel.send(
+    new MessageEmbed()
+    .setColor("RED")
+    .setTitle(`❌ ERROR | An error occurred`)
+    .setDescription(`\`\`\`${e.stack}\`\`\``)
+);
+  }
+  /**
+    * @INFO
+    * Bot Coded by Tomato#6966 | https://github.com/Tomato6966/Discord-Js-Handler-Template
+    * @INFO
+    * Work for Milrato Development | https://milrato.eu
+    * @INFO
+    * Please mention Him / Milrato Development, when using this Code!
+    * @INFO
+  */
+}