+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [1.3.1]
+- Update the Discord.js template
+- Update the Discord.py template
+
 ## [1.3.0]
 - Add the Harmony support